Small refactor
fixed an issue for Devon with testing, and refactored our code to only use one new puzzle base, and moved database stuff into their own folder

Removed current build files within repo, because since it's not editable it didn't update with the new changes.
