+++ MVC/database/__init__.py
@@ -1,0 +1,3 @@
+from .connectDB import *
+from .pangramdb import *
+from .wordstuff import *
