+++ MVC/database/wordstuff.py
@@ -1,7 +1,7 @@
 import sqlite3
 import json
 import os
-from MVC.model import connectDB as cdb
+from MVC.database import connectDB as cdb
 #cursor.execute('''CREATE TABLE dict (word TEXT)''')
 
 
@@ -31,7 +31,7 @@ For the random function, if it's possible to return the word that's randomly cho
 
 #for base game
 def checkWord(userInput):
-    con = cdb.wordBankConnect()
+    con = cdb.connect()
     cursor = con.cursor()
     #sql query to find the word
     cursor.execute("SELECT * FROM dict WHERE word=:word", {'word': userInput})
@@ -57,4 +57,4 @@ def checkWord(userInput):
         con.close()
         return False
 
-checkWord("abdomen")
+#checkWord("abdomen")
