+++ MVC/model/__init__.py
@@ -2,8 +2,5 @@ from .modelrefactor import *
 from .Commands import *
 from .IdentifyBaseWord import *
 from .loadgame import *
-from .pangramdb import *
 from .savegame import *
 from .wordlist import *
-from .wordstuff import *
-from .connectDB import *
