+++ MVC/view/austinguitest.py
@@ -453,7 +453,7 @@ Each puzzle is based off of a pangram, a 7 to 15 letter word that contains 7 uni
         else:
             self.clearInput()
             self.controller.controllerNewGame()
-            self.controller.controllerRunBaseGameGUI(input)
+            self.controller.controllerRunBaseGame(input)
             getLetters = self.controller.controllerGetLetters()
             print("Get letters before: " + getLetters)
             getLetters = getLetters.replace("[", "").replace("]","")
