--- mediatekpackages.egg-info/PKG-INFO
@@ -1,6 +1,0 @@
-Metadata-Version: 2.1
-Name: mediatekpackages
-Version: 1.0
-Author: Devon Fair, Austin An, Tess Hughes, Gabriel Zimmermann, Noah Barger
-Author-email: djfair@millersville.edu, auan@millersville.edu, tmhughe1@millersville.edu, gfzimmer@millersville.edu, nbbarger@millersville.edu
-License-File: LICENSE
