--- mediatekpackages.egg-info/SOURCES.txt
@@ -1,29 +1,0 @@
-LICENSE
-README.md
-setup.py
-MVC/__init__.py
-MVC/test.py
-MVC/controller/__init__.py
-MVC/controller/controllerrefactor.py
-MVC/model/Commands.py
-MVC/model/IdentifyBaseWord.py
-MVC/model/__init__.py
-MVC/model/connectDB.py
-MVC/model/loadgame.py
-MVC/model/modelrefactor.py
-MVC/model/pangramdb.py
-MVC/model/savegame.py
-MVC/model/wordlist.py
-MVC/model/wordstuff.py
-MVC/view/CLIrefactor.py
-MVC/view/__init__.py
-MVC/view/austinguitest.py
-MVC/view/findpath.py
-MVC/view/noahmac.py
-mediatekpackages.egg-info/PKG-INFO
-mediatekpackages.egg-info/SOURCES.txt
-mediatekpackages.egg-info/dependency_links.txt
-mediatekpackages.egg-info/requires.txt
-mediatekpackages.egg-info/top_level.txt
-test/__init__.py
-test/unittests.py
