--- mediatekpackages.egg-info/requires.txt
@@ -1,1 +1,0 @@
-setuptools
